// src/@types/models/create-category-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1.0.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

/**
 *
 *
 * @export
 * @interface CreateCategoryDto
 */
export interface CreateCategoryDto {
  /**
   * @type {Array<Blob>}
   * @memberof CreateProductDto
   */
  icon: Array<Blob>;

  /**
   * @type {string}
   * @memberof CreateCategoryDto
   */
  title: string;

  /**
   * @type {string}
   * @memberof CreateCategoryDto
   */
  parentId?: string;
}